Convert to compiling the PNGs from source
